refactoring run_mrc.py
