+++ code/inference_copy.py
@@ -36,7 +36,8 @@ from trainer_qa import QuestionAnsweringTrainer
 from sparse_retrieval import SparseRetrieval
 from dense_retrieval import DenseRetrieval
 from retrieval_common_part import build_faiss, retrieve_faiss
-from post_processing import post_processing_function
+from postprocessing import post_processing_function
+from run_mrc import run_combine_mrc
 
 from arguments import (
     ModelArguments,
@@ -56,7 +57,7 @@ def main():
     )
     model_args, data_args, training_args = parser.parse_args_into_dataclasses()
 
-    training_args.do_train = True
+    # training_args.do_train = True
 
     print(f"model is from {model_args.model_name_or_path}")
     print(f"data is from {data_args.dataset_name}")
@@ -79,7 +80,7 @@ def main():
 
     # AutoConfig를 이용하여 pretrained model 과 tokenizer를 불러옵니다.
     # argument로 원하는 모델 이름을 설정하면 옵션을 바꿀 수 있습니다.
-    model, tokenizer = cofngiure_model(model_args, training_args, data_args)
+    model, tokenizer = configure_model(model_args, training_args, data_args)
 
     # rue일 경우 : run passage retrieval
     if data_args.eval_retrieval:
@@ -93,7 +94,7 @@ def main():
 
     # eval or predict mrc model
     if training_args.do_eval or training_args.do_predict:
-        run_mrc(data_args, training_args, model_args, datasets, tokenizer, model)
+        run_combine_mrc(data_args, training_args, model_args, datasets, tokenizer, model)
 
 
 def run_retrieval(
@@ -127,47 +128,48 @@ def run_retrieval(
     elif data_args.sparse_name == "elastic":
         df_sparse = retriever_sparse.elastic_retrieve(datasets["validation"], topk=data_args.top_k_retrieval)
 
-    # # 테스트
-    # # Dense Passage Retrieval 부분
-    # retriever_dense = DenseRetrieval(
-    #     tokenize_fn=tokenize_fn, datasets=datasets, data_path=data_path, context_path=context_path 
-    # )
-    # if data_args.dense_name == "None":
-    #     retriever_dense.get_dense_embedding(inbatch=False)
-    #     # 수정
-    #     retrieval_common_part.build_faiss(num_clusters=data_args.num_clusters)
-    #     df_dense = retrieval_common_part.retrieve_faiss(
-    #         datasets["validation"], topk=data_args.top_k_retrieval
-    #     )
-    # elif data_args.dense_name == "in-batch":
-    #     retriever_dense.get_dense_embedding(inbatch=True)
-    #     # 수정
-    #     retrieval_common_part.build_faiss(num_clusters=data_args.num_clusters)
-    #     df_dense = retrieval_common_part.retrieve_faiss(
-    #         datasets["validation"], topk=data_args.top_k_retrieval
-    #     )
+    # 테스트
+    # Dense Passage Retrieval 부분
+    retriever_dense = DenseRetrieval(
+        tokenize_fn=tokenize_fn, datasets=datasets, data_path=data_path, context_path=context_path 
+    )
+    if data_args.dense_name == "None":
+        retriever_dense.get_dense_embedding(inbatch=False)
+        # 수정
+        retrieval_common_part.build_faiss(num_clusters=data_args.num_clusters)
+        df_dense = retrieval_common_part.retrieve_faiss(
+            datasets["validation"], topk=data_args.top_k_retrieval
+        )
+    elif data_args.dense_name == "in-batch":
+        retriever_dense.get_dense_embedding(inbatch=True)
+        # 수정
+        retrieval_common_part.build_faiss(num_clusters=data_args.num_clusters)
+        df_dense = retrieval_common_part.retrieve_faiss(
+            datasets["validation"], topk=data_args.top_k_retrieval
+        )
 
 
     # Sparse Retrieval 결과와 Dense Retrieval 결과를 병합합니다. 
     df = df_sparse
     # 테스트
     for idx in range(len(df_sparse)):
-        # if idx == 10:
-        #     print(df_dense["context"][idx])
-        # temp = df_sparse["context"][idx] + df_dense["context"][idx]
-        # df["context"][idx] = " ".join(temp)
-        df["context"][idx] = " ".join(df_sparse["context"][idx])
+        if idx == 10:
+            print(df_dense["context"][idx])
+        temp = df_sparse["context"][idx] + df_dense["context"][idx]
+        df["context"][idx] = " ".join(temp)
+        # sparse만 이용하는 경우
+        # df["context"][idx] = " ".join(df_sparse["context"][idx])
 
     # Dense Retrieval 결과 일부 출력하기        
-    # df = df_dense
-    # for idx in range(len(df_dense)):
-    #     if idx % 1000 == 0:
-    #         print(df["context_id"][idx])
-    #         print('-----------')
-    #         print(df["question"][idx])
-    #         print('-----------')
-    #         print(df["context"][idx])
-    #     df["context"][idx] = " ".join(df_dense["context"][idx])
+    df = df_dense
+    for idx in range(len(df_dense)):
+        if idx % 1000 == 0:
+            print(df["context_id"][idx])
+            print('-----------')
+            print(df["question"][idx])
+            print('-----------')
+            print(df["context"][idx])
+        df["context"][idx] = " ".join(df_dense["context"][idx])
 
 
     # test data 에 대해선 정답이 없으므로 id question context 로만 데이터셋이 구성됩니다.
@@ -204,86 +206,6 @@ def run_retrieval(
     
     datasets = DatasetDict({"validation": Dataset.from_pandas(df, features=f)})
     return datasets
-
-
-def run_mrc(
-    data_args: DataTrainingArguments,
-    training_args: TrainingArguments,
-    model_args: ModelArguments,
-    datasets: DatasetDict,
-    tokenizer,
-    model,
-) -> NoReturn:
-
-    # eval 혹은 prediction에서만 사용함
-    column_names = datasets["validation"].column_names
-
-    # Padding에 대한 옵션을 설정합니다.
-    # (question|context) 혹은 (context|question)로 세팅 가능합니다.
-
-    # 오류가 있는지 확인합니다.
-    last_checkpoint, max_seq_length = check_no_error(
-        data_args, training_args, datasets, tokenizer
-    )
-
-    # Validation preprocessing / 전처리를 진행합니다.
-
-    eval_dataset = datasets["validation"]
-    prepare_validation_features = preprocess_extract_valid(tokenizer,data_args,column_names,max_seq_length)
-    # Validation Feature 생성
-    eval_dataset = eval_dataset.map(
-        prepare_validation_features,
-        batched=True,
-        num_proc=data_args.preprocessing_num_workers,
-        remove_columns=column_names,
-        load_from_cache_file=not data_args.overwrite_cache,
-    )
-
-    # Data collator
-    # flag가 True이면 이미 max length로 padding된 상태입니다.
-    # 그렇지 않다면 data collator에서 padding을 진행해야합니다.
-    data_collator = DataCollatorWithPadding(
-        tokenizer, pad_to_multiple_of=8 if training_args.fp16 else None
-    )
-
-    metric = load_metric("squad")
-
-    def compute_metrics(p: EvalPrediction) -> Dict:
-        return metric.compute(predictions=p.predictions, references=p.label_ids)
-
-    print("init trainer...")
-    # Trainer 초기화
-    trainer = QuestionAnsweringTrainer(
-        model=model,
-        args=training_args,
-        train_dataset=None,
-        eval_dataset=eval_dataset,
-        eval_examples=datasets["validation"],
-        tokenizer=tokenizer,
-        data_collator=data_collator,
-        post_process_function=post_processing_function,
-        compute_metrics=compute_metrics,
-    )
-
-    logger.info("*** Evaluate ***")
-
-    #### eval dataset & eval example - predictions.json 생성됨
-    if training_args.do_predict:
-        predictions = trainer.predict(
-            test_dataset=eval_dataset, test_examples=datasets["validation"]
-        )
-
-        # predictions.json 은 postprocess_qa_predictions() 호출시 이미 저장됩니다.
-        print(
-            "No metric can be presented because there is no correct answer given. Job done!"
-        )
-
-    if training_args.do_eval:
-        metrics = trainer.evaluate()
-        metrics["eval_samples"] = len(eval_dataset)
-
-        trainer.log_metrics("test", metrics)
-        trainer.save_metrics("test", metrics)
 
 
 if __name__ == "__main__":
