+++ code/train_copy.py
@@ -16,7 +16,7 @@ from transformers import (
 
 from configure import *
 from preprocess import *
-from postprocessing import *
+from postprocessing import post_processing_function, compute_metrics
 
 from utils_qa import postprocess_qa_predictions, check_no_error
 from trainer_qa import QuestionAnsweringTrainer
@@ -25,6 +25,8 @@ from arguments import (
     ModelArguments,
     DataTrainingArguments,
 )
+
+from run_mrc import run_combine_mrc, run_generation_mrc
 
 logger = logging.getLogger(__name__)
 
@@ -55,7 +57,7 @@ def main():
     datasets = load_from_disk(data_args.dataset_name)
     print(datasets)
 
-    model, tokenizer =cofngiure_model(model_args, training_args, data_args)
+    model, tokenizer = configure_model(model_args, training_args, data_args)
     print(
         type(training_args),
         type(model_args),
@@ -67,257 +69,9 @@ def main():
     # do_train mrc model 혹은 do_eval mrc model
     if training_args.do_train or training_args.do_eval:
         if model_args.run_extraction:
-            run_extraction_mrc(data_args, training_args, model_args, datasets, tokenizer, model)
+            run_combine_mrc(data_args, training_args, model_args, datasets, tokenizer, model)
         elif model_args.run_generation:
             run_generation_mrc(data_args, training_args, model_args, datasets, tokenizer, model)
-
-
-def run_extraction_mrc(
-    data_args: DataTrainingArguments,
-    training_args: TrainingArguments,
-    model_args: ModelArguments,
-    datasets: DatasetDict,
-    tokenizer,
-    model,
-) -> NoReturn:
-
-    # dataset을 전처리합니다.
-    # training과 evaluation에서 사용되는 전처리는 아주 조금 다른 형태를 가집니다.
-    if training_args.do_train:
-        column_names = datasets["train"].column_names
-    else:
-        column_names = datasets["validation"].column_names
-
-    # 오류가 있는지 확인합니다.
-    last_checkpoint, max_seq_length = check_no_error(
-        data_args, training_args, datasets, tokenizer
-    )
-
-    if training_args.do_train:
-        if "train" not in datasets:
-            raise ValueError("--do_train requires a train dataset")
-        train_dataset = datasets["train"]
-        
-        # prepare_train_features = preprocess_gen(tokenizer)
-        # dataset에서 train feature를 생성합니다.
-        prepare_train_features = preprocess_extract_train(tokenizer, data_args, column_names, max_seq_length)
-        train_dataset = train_dataset.map(
-            prepare_train_features,
-            batched=True,
-            num_proc=data_args.preprocessing_num_workers,
-            remove_columns=column_names,
-            load_from_cache_file=not data_args.overwrite_cache,
-        )
-
-    if training_args.do_eval:
-        eval_dataset = datasets["validation"]
-
-        prepare_valid_features = preprocess_extract_valid(tokenizer, data_args, column_names, max_seq_length)
-        # Validation Feature 생성
-        eval_dataset = eval_dataset.map(
-            prepare_valid_features,
-            batched=True,
-            num_proc=data_args.preprocessing_num_workers,
-            remove_columns=column_names,
-            load_from_cache_file=not data_args.overwrite_cache,
-        )
-
-    # Data collator
-    # flag가 True이면 이미 max length로 padding된 상태입니다.
-    # 그렇지 않다면 data collator에서 padding을 진행해야합니다.
-    data_collator = DataCollatorWithPadding(
-        tokenizer, pad_to_multiple_of=8 if training_args.fp16 else None
-    )
-
-    # Trainer 초기화
-    trainer = QuestionAnsweringTrainer( 
-        model=model,
-        args=training_args,
-        train_dataset=train_dataset if training_args.do_train else None,
-        eval_dataset=eval_dataset if training_args.do_eval else None,
-        eval_examples=datasets["validation"] if training_args.do_eval else None,
-        tokenizer=tokenizer,
-        data_collator=data_collator,
-        post_process_function=post_processing_function,
-        compute_metrics=compute_metrics,
-    )
-
-    # Training
-    if training_args.do_train:
-        if last_checkpoint is not None:
-            checkpoint = last_checkpoint
-        elif os.path.isdir(model_args.model_name_or_path):
-            checkpoint = model_args.model_name_or_path
-        else:
-            checkpoint = None
-        train_result = trainer.train(resume_from_checkpoint=checkpoint)
-        trainer.save_model()  # Saves the tokenizer too for easy upload
-
-        metrics = train_result.metrics
-        metrics["train_samples"] = len(train_dataset)
-
-        trainer.log_metrics("train", metrics)
-        trainer.save_metrics("train", metrics)
-        trainer.save_state()
-
-        output_train_file = os.path.join(training_args.output_dir, "train_results.txt")
-
-        with open(output_train_file, "w") as writer:
-            logger.info("***** Train results *****")
-            for key, value in sorted(train_result.metrics.items()):
-                logger.info(f"  {key} = {value}")
-                writer.write(f"{key} = {value}\n")
-
-        # State 저장
-        trainer.state.save_to_json(
-            os.path.join(training_args.output_dir, "trainer_state.json")
-        )
-
-    # Evaluation
-    if training_args.do_eval:
-        logger.info("*** Evaluate ***")
-        metrics = trainer.evaluate()
-
-        metrics["eval_samples"] = len(eval_dataset)
-
-        trainer.log_metrics("eval", metrics)
-        trainer.save_metrics("eval", metrics)
-
-
-def run_generation_mrc(
-    data_args: DataTrainingArguments,
-    training_args: TrainingArguments,
-    model_args: ModelArguments,
-    datasets: DatasetDict,
-    tokenizer,
-    model,
-) -> NoReturn:
-
-    # Training Arguments를 Seq2Seq로 바꿔준다. 이 때, predict_with_generate만 True로 변경해주고 나머지는 Default 사용
-    training_args = Seq2SeqTrainingArguments(
-        predict_with_generate=True
-    )
-
-    # dataset을 전처리합니다.
-    # training과 evaluation에서 사용되는 전처리는 아주 조금 다른 형태를 가집니다.
-    if training_args.do_train:
-        column_names = datasets["train"].column_names
-    else:
-        column_names = datasets["validation"].column_names
-    
-    print(data_args.pad_to_max_length)
-    print(tokenizer)
-
-    max_seq_length = data_args.max_seq_length
-    max_length = data_args.max_answer_length
-
-    if training_args.do_train:
-        if "train" not in datasets:
-            raise ValueError("--do_train requires a train dataset")
-
-        train_dataset = datasets["train"]
-
-        preprocess_function = preprocess_gen(tokenizer)
-        train_dataset = train_dataset.map(
-            preprocess_function,
-            batched=True,
-            num_proc=data_args.preprocessing_num_workers,
-            remove_columns=column_names,
-            load_from_cache_file=False,
-        )
-
-    if training_args.do_eval:
-        eval_dataset = datasets["validation"]
-
-        eval_dataset = eval_dataset.map(
-            preprocess_function,
-            batched=True,
-            num_proc=data_args.preprocessing_num_workers,
-            remove_columns=column_names,
-            load_from_cache_file=False,
-        )
-
-    data_collator = DataCollatorWithPadding(
-        tokenizer, 
-        pad_to_multiple_of=8 if training_args.fp16 else None
-    )
-
-    def postprocess_text(preds, labels):
-        preds = [pred.strip() for pred in preds]
-        labels = [label.strip() for label in labels]
-        
-        preds = ["\n".join(tokenizer.tokenize(pred)) for pred in preds]
-        labels = ["\n".join(tokenizer.tokenize(label)) for label in labels]
-
-        return preds, labels
-
-
-    metric = load_metric("squad")
-
-    def compute_metrics(eval_preds):
-        preds, labels = eval_preds
-
-        if isinstance(preds, tuple):
-            preds = preds[0]
-
-        max_val_samples = 16
-
-        decoded_preds = tokenizer.batch_decode(preds, skip_special_tokens=True)
-        decoded_labels = tokenizer.batch_decode(labels, skip_special_tokens=True)
-
-        decoded_preds, decoded_labels = postprocess_text(decoded_preds, decoded_labels)
-
-        formatted_predictions = [{"id": ex["id"], "prediction_text": decoded_preds[i]} for i, ex in         
-                                 enumerate(datasets["validation"].select(range(max_val_samples)))]
-        references = [{"id": ex["id"], "answers": ex["answers"]} for ex in datasets["validation"].select(range(max_val_samples))]
-
-        result = metric.compute(predictions=formatted_predictions, references=references)
-
-        return result
-    
-
-    trainer = Seq2SeqTrainer(
-            model=model,
-            args=training_args,
-            train_dataset=train_dataset,
-            eval_dataset=eval_dataset,
-            tokenizer=tokenizer,
-            data_collator=data_collator,
-            compute_metrics=compute_metrics,   
-        )
-
-    if training_args.do_train:
-        train_result = trainer.train()
-        trainer.save_model()  # Saves the tokenizer too for easy upload
-
-        metrics = train_result.metrics
-        metrics["train_samples"] = len(train_dataset)
-
-        trainer.log_metrics("train", metrics)
-        trainer.save_metrics("train", metrics)
-        trainer.save_state()
-
-        output_train_file = os.path.join(training_args.output_dir, "train_results.txt")
-
-        with open(output_train_file, "w") as writer:
-            logger.info("***** Train results *****")
-            for key, value in sorted(train_result.metrics.items()):
-                logger.info(f"  {key} = {value}")
-                writer.write(f"{key} = {value}\n")
-
-        trainer.state.save_to_json(
-            os.path.join(training_args.output_dir, "trainer_state.json")
-        )
-
-    # Evaluation
-    if training_args.do_eval:
-        logger.info("*** Evaluate ***")
-        metrics = trainer.evaluate()
-
-        metrics["eval_samples"] = len(eval_dataset)
-
-        trainer.log_metrics("eval", metrics)
-        trainer.save_metrics("eval", metrics)
 
 
 if __name__ == "__main__":
